refactor: remove any type props

+++ frontend/src/modal/ReturnModal.tsx
@@ -1,9 +1,20 @@
 // import axios from "axios";
 import { axiosReturn } from "../network/axios/axios.custom";
 import "./returnModal.css";
+// import { userInfo } from "../types/userTypes";
 // import { useState } from "react";
 
-export default function ReturnModal(props: any) {
+import { lentCabinetInfo } from "../redux/slices/lentSlice";
+import { userInfo } from "../redux/slices/userSlice";
+
+interface ReturnModalProps {
+  user: userInfo | undefined;
+  lentCabinet: lentCabinetInfo | undefined;
+  setContent: React.Dispatch<React.SetStateAction<string>>;
+  setPath: React.Dispatch<React.SetStateAction<string>>;
+}
+
+export default function ReturnModal(props: ReturnModalProps) {
 
   // 리턴 모달에 이벤트를 띄울 시, 하단의 함수를 사용하여 returnTarget의 값을 바꾸어 사용했어요.
   // 현재 리턴모달 측 변경하는 이벤트를 사용하지 않으므로, returnTarget의 값은 기본값인 contentsmodal 이 됩니다.
@@ -57,17 +68,20 @@ export default function ReturnModal(props: any) {
     //     props.setContent("다시 시도해주세요!");
     //     props.setPath("");
     //   });
-    axiosReturn(props.lentCabinet.lent_id)
-      .then((res: any) => {
-        localStorage.clear();
-        props.setContent("반납되었습니다");
-        props.setPath("/lent");
-      })
-      .catch((err: any) => {
-        console.error(err);
-        props.setContent("다시 시도해주세요!");
-        props.setPath("");
-      });
+    if (props.lentCabinet)
+    {
+      axiosReturn(props.lentCabinet.lent_id)
+        .then((res: any) => {
+          localStorage.clear();
+          props.setContent("반납되었습니다");
+          props.setPath("/lent");
+        })
+        .catch((err: any) => {
+          console.error(err);
+          props.setContent("다시 시도해주세요!");
+          props.setPath("");
+        });
+    }
       // 5월 이벤트 사용을 위해, event 테이블에 당첨자가 return 했음을 알려주는 함수입니다.
       // handleEventReturn();
   };
